better export-www and removed counter files

+++ export-www.sh
@@ -1,18 +1,18 @@
 #!/bin/bash
 
-# exit if an error occurs
-set -e
-
 WWWDIR=/www/Abt-Optimization/Personal/miltenberger
 
-#echo "pushing repository..."
-#git push
+echo "pushing repository..."
+if ! git push
+  exit
 
 echo "exporting..."
-git archive master | tar -x -C $WWWDIR
+if ! git archive master | tar -x -C $WWWDIR
+  exit
 
 echo "setting permissions..."
 cd $WWWDIR
 rm *.sh
+rm *.py
 rm .gitignore
-chmod -R a+rX,ug+w *
+chmod -f -R a+rX,ug+w *
